@@ -2,38 +2,39 @@
     "name": "02-inversify-example",
     "version": "1.0.0",
     "description": "",
-    "main": "index.js",
+    "main": "src/handler.js",
     "scripts": {
         "start": "tsc && node dist/handler.js",
         "lint": "npx eslint src/ --ext .js,.jsx,.ts,.tsx",
         "lint-fix": "npx eslint src/ --ext .js,.jsx,.ts,.tsx --fix",
         "test": "jest",
+        "test:watch": "jest --watchAll --coverage",
         "build": "rm -rf dist && tsc"
     },
     "author": "joaosalless@gmail.com",
     "license": "ISC",
     "dependencies": {
-        "typescript": "^3.8.3",
+        "@aws-sdk/client-api-gateway": "^3.56.0",
         "inversify": "^6.0.1",
         "reflect-metadata": "^0.1.13"
     },
     "devDependencies": {
-        "@typescript-eslint/parser": "^2.31.0",
-        "@typescript-eslint/eslint-plugin": "^4.14.2",
-        "@types/jest": "26.0.20",
-        "@types/node": "^13.13.12",
+        "@types/jest": "^27.4.1",
+        "@types/node": "^17.0.23",
         "@types/request-promise": "^4.1.46",
         "@types/source-map-support": "^0.5.3",
         "@types/validator": "^13.1.3",
+        "@typescript-eslint/eslint-plugin": "^4.14.2",
+        "@typescript-eslint/parser": "^2.31.0",
+        "eslint": "^6.8.0",
         "eslint-config-airbnb-typescript": "^12.0.0",
         "eslint-plugin-import": "^2.22.1",
-        "jest": "26.0.1",
+        "jest": "^27.5.1",
         "onchange": "^7.1.0",
         "serve": "^11.3.2",
         "source-map-support": "^0.5.19",
-        "ts-jest": "26.0.0",
-        "ts-node": "^8.10.2",
-        "typescript": "^3.9.5",
-        "eslint": "^6.8.0"
+        "ts-jest": "^27.1.4",
+        "ts-node": "^10.7.0",
+        "typescript": "^4.6.3"
     }
 }
