@@ -1,38 +1,38 @@
 import "reflect-metadata";
-import TYPES from "interfaces/types";
-import { ConfigServiceInterface } from './config-service';
-import { inject, injectable } from "inversify";
+import { TYPES } from '../interfaces/container';
+import { AbstractService, AbstractServiceInterface } from "./abstract-service";
+import { Container, inject, injectable, interfaces } from "inversify";
 
-export interface LoggerServiceInterface {
-  debug(message: string, data: object);
-  info(message: string, data: object);
-  warning(message: string, data: object);
-  error(message: string, data: object);
+export interface LoggerServiceInterface extends AbstractServiceInterface {
+  debug(message: string, data?: object);
+  info(message: string, data?: object);
+  warning(message: string, data?: object);
+  error(message: string, data?: object);
 }
 
 @injectable()
-export class LoggerService implements LoggerServiceInterface {
-  constructor(
-    @inject(TYPES.ConfigServiceInterface) private config: ConfigServiceInterface,
-  ) {}
+export class LoggerService extends AbstractService implements LoggerServiceInterface {
+  constructor(@inject(TYPES.Container) protected _container: interfaces.Container) {
+    super(_container);
+  }
 
   log(message: string, data: object, level: string) {
-    console.log(message, data, level);
+    console.log(`${level}:`, message, data);
   }
 
-  debug(message: string, data: object) {
-    this.log(message, data, 'debug');
+  debug(message: string, data?: object) {
+    this.log(message, data || {}, 'debug');
   }
-  
-  info(message: string, data: object) {
-    this.log(message, data, 'info');
+
+  info(message: string, data?: object) {
+    this.log(message, data || {}, 'info');
   }
-  
-  warning(message: string, data: object) {
-    this.log(message, data, 'warning');
+
+  warning(message: string, data?: object) {
+    this.log(message, data || {}, 'warning');
   }
-  
-  error(message: string, data: object) {
-    this.log(message, data, 'error');
+
+  error(message: string, data?: object) {
+    this.log(message, data || {}, 'error');
   }
 }
