@@ -1,60 +1,102 @@
 import "reflect-metadata";
-import TYPES from "interfaces/types";
-import { JsonSchemaServiceInterface } from './json-schema-service';
+import { TYPES } from '../interfaces/container';
 import { ConfigServiceInterface } from './config-service';
-import { LoggerServiceInterface } from './logger-service';
 import { I18nServiceInterface } from './i18n-service';
-import { inject, injectable } from "inversify";
+import { inject, injectable, interfaces } from "inversify";
+import { JsonSchemaServiceInterface } from './json-schema-service';
+import { LoggerServiceInterface } from './logger-service';
+import { AbstractService, AbstractServiceInterface } from "./abstract-service";
 
-export interface HttpServiceInterface {
-  request(params: { method: string, url: string, body: any, headers: any }): any;
-  response(params: { code: number, data: any, error: any }): any;
+export interface HttpRequestPayloadInterface {
+  method: string; 
+  url: string; 
+  body: any; 
+  headers: any; 
+  schema?: string;
+}
+
+export interface HttpResponsePayloadInterface {
+  code: number;
+  data?: any;
+  error?: any;
+  schema?: any;
+}
+
+export interface HttpResponseErrorPayloadInterface {
+  code: number;
+  error?: any;
+  schema?: any;
+}
+
+export interface HttpResponseInterface {
+  code: number;
+  data?: any;
+  message?: any;
+}
+
+export interface HttpServiceInterface extends AbstractServiceInterface {
+  request(params: HttpRequestPayloadInterface): Promise<HttpResponseInterface>;
+  response(params: HttpResponsePayloadInterface): Promise<HttpResponseInterface>;
+  responseError(params: HttpResponseErrorPayloadInterface): Promise<HttpResponseInterface>;
 }
 
 @injectable()
-export class HttpService implements HttpServiceInterface {
-  constructor(
-    @inject(TYPES.LoggerServiceInterface) private logger: LoggerServiceInterface,
-    @inject(TYPES.ConfigServiceInterface) private config: ConfigServiceInterface,
-    @inject(TYPES.I18nServiceInterface) private i18n: I18nServiceInterface,
-    @inject(TYPES.JsonSchemaServiceInterface) private jsonSchema: JsonSchemaServiceInterface,
-  ) {}
+export class HttpService extends AbstractService implements HttpServiceInterface {
+  private logger: LoggerServiceInterface;
+  private config: ConfigServiceInterface;
+  private i18n: I18nServiceInterface;
+  private jsonSchema: JsonSchemaServiceInterface;
 
-  request(params: { method: string; url: string; body: any; headers: any; schemaId?: string; }) {
+  constructor(@inject(TYPES.Container) protected _container: interfaces.Container) {
+    super(_container);
+    this.logger = _container.get<LoggerServiceInterface>(TYPES.LoggerServiceInterface);
+    this.config = _container.get<ConfigServiceInterface>(TYPES.ConfigServiceInterface);
+    this.i18n = _container.get<I18nServiceInterface>(TYPES.I18nServiceInterface);
+    this.jsonSchema = _container.get<JsonSchemaServiceInterface>(TYPES.JsonSchemaServiceInterface);
+  }
+
+  public async request(params: HttpRequestPayloadInterface): Promise<HttpResponseInterface> {
+    const { method, url, body, headers, schema } = params;
+
     try {
-      this.jsonSchema.validate(schemaId, body);
-
-      //
+      return Promise.resolve({
+        code: 500,
+        message: 'Not implemented.',
+      });
     } catch (error) {
-      this.logger.error('HttpService.request error', { error })
+      this.responseError({ code: 500, error });
     }
   }
 
-  response(params: { code: number; data: any; error?: any; schema?: string; }) {
+  public async response(params: HttpResponsePayloadInterface): Promise<HttpResponseInterface> {
+    const { code, data, error, schema } = params;
+
     try {
+      const response: any = {
+        code,
+        data,
+      };
       this.jsonSchema.validate(schema, data);
 
-      //
+      return Promise.resolve(response);
     } catch (error) {
-      this.logger.error('HttpService.request error', { error })
+      this.responseError({ code: 500, error });
     }
   }
 
-  responseError(params: { code: number; data: any; error?: any; schema?: string; }) {
-    try {
-      this.jsonSchema.validate(schema, data);
+  public async responseError(params: HttpResponseErrorPayloadInterface): Promise<HttpResponseInterface> {
+    const { code, error, schema } = params;
 
-      //
+    const response: any = {
+      code,
+      message: error?.message,
+    };
+
+    try {
+      this.jsonSchema.validate(schema, response);
+      return Promise.resolve(response);
     } catch (error) {
       this.logger.error('HttpService.request error', { error })
     }
-
-    const createResponse = (statusCode: number, body: any): Promise<any> => Promise.resolve({
-      statusCode,
-      headers: {
-        'Access-Control-Allow-Origin': '*',
-      },
-      body: JSON.stringify(body),
-    });
   }
 }
